requirements.txt
fastapi
langchain
langchain_ollama
langchain-openai
langchain-anthropic
langchain-mcp-adapters

mcp[cli]

# Core dependencies
pydantic>=2.0.0
httpx>=0.24.0
tiktoken>=0.5.0
PyYAML>=6.0

# Async support
anyio>=3.6.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Logging and monitoring
structlog>=23.0.0

# Optional: For enhanced LLM support
openai>=1.0.0
anthropic>=0.3.0

# Optional: For advanced caching
redis>=4.5.0